requirements.txt
pytest==8.2.1
pydantic==2.7.1
requests==2.32.1
allure-pytest==2.13.5
python-dotenv==1.0.1